Added support for Rate of Turn using the built-in IMU chip on the Arduino Nano 33 IoT board. It collects the Z axis from the gyroscope and sends that over bluetooth via a new Characteristic. This still needs to be tested on the Arduino Nano 33 BLE, which has a different IMU chip (the BLE also has a magnetometer that the IoT does not). The code may still work ... just need to test it out.
The idea of this update  is to be able to include a turn coordinator on the
AoA Indicator screen (Android app) that shows the rate of turn (e.g. degrees
per second, which can be translated to display like the tiny airplane in a
standard turn coordinator).

At some point, it might be interesting to include the Accelerometer data
from the IMU that could display slip/skid data similar to the ball in a
standard turn coordinator.